// apps/backend/src/modules/events/enums/event-type.enum.ts
export enum EventType {
  LANDING_PAGE_VIEWED = 'LANDING_PAGE_VIEWED',
  LOAN_PAGE_VIEWED = 'LOAN_PAGE_VIEWED',
  EMI_CALCULATOR_USED = 'EMI_CALCULATOR_USED',
  SALARY_ENTERED = 'SALARY_ENTERED',
  OTP_VERIFIED = 'OTP_VERIFIED',
  FORM_STARTED = 'FORM_STARTED',
  FORM_ABANDONED = 'FORM_ABANDONED',
  DOCUMENT_UPLOADED = 'DOCUMENT_UPLOADED',
  CALLBACK_REQUESTED = 'CALLBACK_REQUESTED',
  LEAD_INACTIVE = "LEAD_INACTIVE",
}